requirement.txt
astroid==2.0.4
colorama==0.3.9
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.1.1
PyYAML==3.13
requests==2.20.1
wrapt==1.10.11
